+++ src/main/java/fuzs/configmenusforge/ConfigMenusForge.java
@@ -1,12 +1,8 @@
 package fuzs.configmenusforge;
 
 import fuzs.configmenusforge.config.TestConfig;
-import fuzs.configmenusforge.lib.core.EnvTypeExecutor;
 import fuzs.configmenusforge.lib.core.ModLoaderEnvironment;
 import fuzs.configmenusforge.lib.network.NetworkHandler;
-import fuzs.configmenusforge.lib.proxy.ClientProxy;
-import fuzs.configmenusforge.lib.proxy.IProxy;
-import fuzs.configmenusforge.lib.proxy.ServerProxy;
 import fuzs.configmenusforge.network.client.message.C2SAskPermissionsMessage;
 import fuzs.configmenusforge.network.client.message.C2SSendConfigMessage;
 import fuzs.configmenusforge.network.message.S2CGrantPermissionsMessage;
@@ -30,12 +26,11 @@ import java.io.File;
 public class ConfigMenusForge {
 
     public static final String MOD_ID = "configmenusforge";
-    public static final String NAME = "Config Menus for Forge";
-    public static final String URL = "https://www.curseforge.com/minecraft/mc-mods/config-menus-forge";
-    public static final Logger LOGGER = LogManager.getLogger(ConfigMenusForge.NAME);
+    public static final String MOD_NAME = "Config Menus for Forge";
+    public static final String MOD_URL = "https://www.curseforge.com/minecraft/mc-mods/config-menus-forge";
+    public static final Logger LOGGER = LogManager.getLogger(ConfigMenusForge.MOD_NAME);
 
-    @SuppressWarnings("Convert2MethodRef")
-    public static final IProxy PROXY = EnvTypeExecutor.runForDist(() -> () -> new ClientProxy(), () -> () -> new ServerProxy());
+    public static final NetworkHandler NETWORK = NetworkHandler.of(MOD_ID, true, false);
 
     @SubscribeEvent
     public static void onConstructMod(final FMLConstructModEvent evt) {
@@ -45,10 +40,10 @@ public class ConfigMenusForge {
     }
 
     private static void registerMessages() {
-        NetworkHandler.INSTANCE.register(C2SAskPermissionsMessage.class, C2SAskPermissionsMessage::new, NetworkDirection.PLAY_TO_SERVER);
-        NetworkHandler.INSTANCE.register(S2CGrantPermissionsMessage.class, S2CGrantPermissionsMessage::new, NetworkDirection.PLAY_TO_CLIENT);
-        NetworkHandler.INSTANCE.register(C2SSendConfigMessage.class, C2SSendConfigMessage::new, NetworkDirection.PLAY_TO_SERVER);
-        NetworkHandler.INSTANCE.register(S2CUpdateConfigMessage.class, S2CUpdateConfigMessage::new, NetworkDirection.PLAY_TO_CLIENT);
+        NETWORK.register(C2SAskPermissionsMessage.class, C2SAskPermissionsMessage::new, NetworkDirection.PLAY_TO_SERVER);
+        NETWORK.register(S2CGrantPermissionsMessage.class, S2CGrantPermissionsMessage::new, NetworkDirection.PLAY_TO_CLIENT);
+        NETWORK.register(C2SSendConfigMessage.class, C2SSendConfigMessage::new, NetworkDirection.PLAY_TO_SERVER);
+        NETWORK.register(S2CUpdateConfigMessage.class, S2CUpdateConfigMessage::new, NetworkDirection.PLAY_TO_CLIENT);
     }
 
     private static void addTestConfigs() {