+++ src/main/java/fuzs/configmenusforge/client/gui/screens/ConfigScreen.java
@@ -117,7 +117,6 @@ public abstract class ConfigScreen extends Screen {
     }
 
     private static class Main extends ConfigScreen {
-
         /**
          * called when closing screen via done button
          */
@@ -135,8 +134,11 @@ public abstract class ConfigScreen extends Screen {
         protected void init() {
             super.init();
             this.doneButton = this.addRenderableWidget(new Button(this.width / 2 - 154, this.height - 28, 150, 20, CommonComponents.GUI_DONE, button -> {
-                this.valueToData.values().forEach(IEntryData::saveConfigValue);
-                this.onSave.run();
+                // avoid unnecessary config reloading
+                if (this.valueToData.values().stream().anyMatch(Predicate.not(IEntryData::mayDiscardChanges))) {
+                    this.valueToData.values().forEach(IEntryData::saveConfigValue);
+                    this.onSave.run();
+                }
                 this.minecraft.setScreen(this.lastScreen);
             }));
             this.cancelButton = this.addRenderableWidget(new Button(this.width / 2 + 4, this.height - 28, 150, 20, CommonComponents.GUI_CANCEL, button -> this.onClose()));