+++ src/main/java/fuzs/configmenusforge/client/gui/util/ScreenUtil.java
@@ -27,7 +27,7 @@ import java.util.stream.Stream;
 public class ScreenUtil {
     private static final ResourceLocation TINY_JUMPER_LOCATION = new ResourceLocation(ConfigMenusForge.MOD_ID, "textures/gui/tiny_jumper.png");
     private static final Component INFO_1_TOOLTIP = new TranslatableComponent("configmenusforge.gui.info.1");
-    private static final Component INFO_2_TOOLTIP = new TranslatableComponent("configmenusforge.gui.info.2", ConfigMenusForge.NAME);
+    private static final Component INFO_2_TOOLTIP = new TranslatableComponent("configmenusforge.gui.info.2", ConfigMenusForge.MOD_NAME);
     private static final Component INFO_3_TOOLTIP = new TranslatableComponent("configmenusforge.gui.info.3");
 
     public static ConfirmScreen makeConfirmationScreen(BooleanConsumer booleanConsumer, Component component1, Component component2, ResourceLocation background) {
@@ -54,7 +54,7 @@ public class ScreenUtil {
 
     public static AnimatedIconButton makeModPageButton(int posX, int posY, Font font, Consumer<Style> handleComponentClicked, ITooltipRenderer renderTooltip) {
         return new AnimatedIconButton(posX, posY, 20, 20, 0, 0, TINY_JUMPER_LOCATION, button -> {
-            Style style = Style.EMPTY.withClickEvent(new ClickEvent(ClickEvent.Action.OPEN_URL, ConfigMenusForge.URL));
+            Style style = Style.EMPTY.withClickEvent(new ClickEvent(ClickEvent.Action.OPEN_URL, ConfigMenusForge.MOD_URL));
             handleComponentClicked.accept(style);
         }, (Button button, PoseStack poseStack, int mouseX, int mouseY) -> {
             final List<FormattedCharSequence> tooltip = Stream.of(INFO_1_TOOLTIP, INFO_2_TOOLTIP, INFO_3_TOOLTIP)