+++ src/main/java/fuzs/configmenusforge/client/util/ServerConfigUploader.java
@@ -1,7 +1,8 @@
 package fuzs.configmenusforge.client.util;
 
 import com.electronwill.nightconfig.toml.TomlFormat;
-import fuzs.configmenusforge.lib.network.NetworkHandler;
+import fuzs.configmenusforge.ConfigMenusForge;
+import fuzs.configmenusforge.lib.core.ReflectionHelper;
 import fuzs.configmenusforge.network.client.message.C2SSendConfigMessage;
 import net.minecraft.client.Minecraft;
 import net.minecraftforge.common.ForgeConfigSpec;
@@ -13,10 +14,11 @@ public class ServerConfigUploader {
 
     public static void saveAndUpload(ModConfig config) {
         ((ForgeConfigSpec) config.getSpec()).save();
+        ModConfigSync.fireReloadingEvent(config);
         if (config.getType() == ModConfig.Type.SERVER && !Minecraft.getInstance().isLocalServer()) {
             final ByteArrayOutputStream stream = new ByteArrayOutputStream();
             TomlFormat.instance().createWriter().write(config.getConfigData(), stream);
-            NetworkHandler.INSTANCE.sendToServer(new C2SSendConfigMessage(config.getFileName(), stream.toByteArray()));
+            ConfigMenusForge.NETWORK.sendToServer(new C2SSendConfigMessage(config.getFileName(), stream.toByteArray()));
         }
     }
 }