+++ CHANGELOG.md
@@ -3,7 +3,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog].
 
-## [v2.0-1.16.5] - 2021-11-05
+## [v2.0.1-1.17.1] - 2021-11-07
+### Changed
+- Configs are now reloaded immediately after being edited
+### Fixed
+- Fixed vanilla servers showing as incompatible when they aren't
+- Fixed server config reloading event not being fired for the player that changed the config
+
+## [v2.0.0-1.17.1] - 2021-11-05
 - Ported to Minecraft 1.17
 
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/